requirements.txt
asgiref==3.7.2
Django~=5.0.4
sqlparse==0.5.0
Pillow==10.3.0
flower==2.0.1
celery==5.4.0
stripe==9.3.0
python-decouple==3.8
WeasyPrint==61.2

django-import-export~=4.3.7